modified data from app .jsx and moved it to the correct components, I also tried playing  around with the form.
