+++ src/App.jsx
@@ -1,12 +1,12 @@
 import { useState } from "react";
 import eventsData from "./data";
-// import { v1 as generateUniqueID } from "uuid";
-// import Attendees from "./Attendees";
+//  import { v1 as generateUniqueID } from "uuid";
+  // import Attendees from "./Components/Attendees.jsx"
 import Event from "./Components/Event.jsx";
 import Footer from "./Components/Footer.jsx";
 import Header from "./Components/Header.jsx";
 import NewEventForm from "./Components/NewEventForm.jsx";
-// import Attendee from "./Components/Attendee"
+  // import Attendee from "./Components/Attendee.jsx"
 
 
 
@@ -20,15 +20,15 @@ function App() {
 
   function updateEventAttendance(eventId,attendeeId) {
     const eventArray =[...events]
-    const eventIndex =eventArray.findIndex((event)=> eventId === eventId)
-    const event ={...eventArray[eventIndex] }
+    const eventIndex =eventArray.findIndex((event)=> eventId === event.id)
+    const event = {...eventArray[eventIndex] }
     const personIndex = event.people.findIndex(
       (person)=> person.id === attendeeId
     )
     const peopleArray = [...event.people]
     peopleArray[personIndex].attendance = !peopleArray[personIndex].attendance
     event.people = peopleArray
-    eventArray[eventIndex] =event
+    eventArray[eventIndex] = event
     setEvents(eventArray)
   }
 
@@ -45,12 +45,17 @@ function App() {
   </div>
   <div className="events">
     <ul>
+ 
+
+
       {events.map((event)=>(
+
         <Event
         event={event}
         updateEventAttendance={updateEventAttendance} />
       ))}
     </ul>
+
   </div>
 </main>
 <Footer />
@@ -61,121 +66,18 @@ function App() {
 
 export default App
 
-  const [selectOption, setSelectOption] = useState("");
-
-  const [newEvent, setNewEvent] = useState({
-    id: "",
-    eventType: "",
-    name: "",
-    organizer: "",
-    eventImage: "",
-    date: "",
-    people: [],
-  });
-
-  function addEvent() {
-    const createEvent = {
-      id: generateUniqueID(),
-      eventType: selectOption,
-      name: newEvent.name,
-      organizer: newEvent.organizer,
-      eventImage: newEvent.eventImage || "https://loremflickr.com/640/480/",
-      date: newEvent.date,
-      people: [],
-    };
-    handleAddEvent(createEvent);
-  }
-
-  function handleSelectChange(e) {
-    setSelectOption(e.target.value);
-  }
-
-  function handleSubmit(e) {
-    e.preventDefault();
-    addEvent();
-    resetEventForm();
-  }
-
-  function handleTextChange(e) {
-    setNewEvent({
-      ...newEvent,
-      [e.target.id]: e.target.value,
-    });
-  }
-
-  function resetEventForm() {
-    setNewEvent({
-      id: "",
-      eventType: "",
-      name: "",
-      organizer: "",
-      eventImage: "",
-      date: "",
-    });
-    setSelectOption("");
-  }
-
-  function handleAddEvent(event) {
-    setEvents([event, ...events]);
-  }
-
-  function toggleEventAttendees() {
-    setShowAttendees(!showAttendees);
-  }
+  
 
   
+
+  
+
+ 
 
 
   
         
-      </>
-      <main>
-        <div className="new-event">
-          <>
-            <form onSubmit={handleSubmit}>
-              <h3>Create a new event</h3>
-              <label htmlFor="name">Event name:</label>
-              <input
-                type="text"
-                id="name"
-                onChange={handleTextChange}
-                value={newEvent.name}
-              />
-
-              <label htmlFor="organizer">Organizer:</label>
-              <input
-                type="text"
-                id="organizer"
-                onChange={handleTextChange}
-                value={newEvent.organizer}
-              />
-
-              <label htmlFor="eventImage">Event image:</label>
-              <input
-                type="text"
-                id="eventImage"
-                onChange={handleTextChange}
-                value={newEvent.eventImage}
-              />
-              <label htmlFor="eventType">Event type:</label>
-              <select id="eventType" onChange={handleSelectChange}>
-                <option value=""></option>
-                <option value="Birthday">Birthday</option>
-                <option value="Anniversary">Anniversary</option>
-                <option value="Intramural Sport">Intramural Sport</option>
-                <option value="Watch Party">Watch Party</option>
-                <option value="wedding">Wedding</option>
-              </select>
-              <br />
-              <input type="submit" />
-            </form>
-          </>
-        </div>
-        <div className="events">
-          <ul>
-            {events.map((event) => {
-              const { people: attendees } = event;
-
+
